Fix warnings of Dart Analyzer
* Disabled LocalStorage tests for now, need to rethink its approach
* Refactored MapDirty, now it is HashMapDirty, and doesn't use mirrors
* Fixed a bunch of warnings in Dart Editor

Now, there are only 3 warnings left, but I think we can leave them.
They are:

* The return type 'HttpRequestMock' is not a 'HttpRequestMaybeMock'
* The return type 'HttpRequest' is not a 'HttpRequestMaybeMock'

This is difficult to avoid, since we intentionally return not
HttpRequestMaybeMock, to be able to mock HTTP requests.

And also:

* Missing inherited members: 'Request.get', 'Request.delete',
  'Request.put' and 'Request.post'

Not sure how to avoid this as well, since we have to implement Request,
but extend from Mock and actually mock these methods.

I guess we can live with these for now.